requirements.txt
websockets==10.4
openai==0.27.1
python-decouple==3.8
pydub==0.25.1
ffmpeg==1.4
backoff==2.2.1